Moving HeapBucketStats into its own file
