+++ lib/Common/Memory/HeapBlock.h
@@ -2,6 +2,11 @@
 // Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
+
+#pragma once
+
+#include "HeapBucketStats.h"
+
 class ScriptMemoryDumper;
 
 namespace Memory
@@ -39,36 +44,6 @@ class  RecyclerSweep;
 class MarkContext;
 
 #if ENABLE_MEM_STATS
-struct MemStats
-{
-    size_t objectByteCount;
-    size_t totalByteCount;
-
-    MemStats();
-
-    void Reset();
-    size_t FreeBytes() const;
-    double UsedRatio() const;
-    void Aggregate(const MemStats& other);
-};
-
-struct HeapBucketStats: MemStats
-{
-#ifdef DUMP_FRAGMENTATION_STATS
-    uint totalBlockCount;
-    uint objectCount;
-    uint finalizeCount;
-
-    HeapBucketStats();
-    void Reset();
-    void Dump() const;
-#endif
-
-    void PreAggregate();
-    void BeginAggregate();
-    void Aggregate(const HeapBucketStats& other);
-};
-
 #ifdef DUMP_FRAGMENTATION_STATS
 #define DUMP_FRAGMENTATION_STATS_ONLY(x) x
 #define DUMP_FRAGMENTATION_STATS_IS(x) x
